refactor: the ctx and the stdlib cache each split along their own subjects
Both files sat past the reading budget, which is what stopped either carrying
an audit stamp. Each held more than one subject, so each splits rather than
argues for an exception.

`primitives/ctx.rs` keeps the two capabilities — the pure `Alloc` and the
VM-bearing `NativeCtx` — and hands the ergonomic `ctx.*` constructor surface to
`ctx/build.rs` and the throwaway-VM test seams to `ctx/scaffold.rs`. The seams
keep their `primitives::ctx::` path, which is what the external `tests/` crates
name them by.

`compiler/stdlib_cache.rs` keeps the implementation and splits its tests along
the two subjects they already had: what the serialized form carries across
(`tests/format.rs`) and how the file on disk behaves (`tests/file.rs`).

The audit that follows: both module headers restated their document rather than
citing it, the entry constant pool was described as holding live closures where
it holds none, and the pruning comment quoted a file size twice what a store
writes.
