+++ src/compiler/stdlib_cache.rs
@@ -1,31 +1,7 @@
-//! Disk cache for the standard-library compilation.
-//!
-//! `init_stdlib` recompiles stdlib.lisp (~2850 lines) on every process start,
-//! costing ~2.4s in `compile_file` before execution (5ms) even runs. The whole
-//! front end (expand → analyze → regions → lower → emit) is deterministic:
-//! same stdlib source, same elle binary → same bytecode. This module turns that
-//! work into a one-time cost by serializing the compiled `Bytecode` (plus the
-//! per-closure `ClosureTemplate`s and their LIR, so the JIT keeps working) to a
-//! content-addressed cache file, keyed by the running binary's identity +
-//! stdlib source hash + the canonical primitive-table identity (the last
-//! because serialized native-fn immediates carry process-local `prim_id`s).
-//!
-//! Serialization strategy: the cache format is a plain `StoredBytecode` struct
-//! that is 100% owned data — no `Rc`, no pointers. A symbol or keyword id is
-//! its name's hash (docs/impl/symbol.md), the same number in every process, so
-//! the ids travel as they stand; only the spellings ride alongside, in the
-//! `names` table, and replay into the loading instance's display memo.
-//! `Value`s that appear in the constant pool are scalars
-//! (int/float/bool/nil/keyword/symbol) by construction — string and compound
-//! literals lower to `MaterializeConst` templates, not pool constants — so the
-//! pool serializes cheaply. Closures recurse via `child_protos`.
-//!
-//! LIR: the JIT compiles from `ClosureTemplate.lir_function` in the background.
-//! If the cache dropped LIR, every stdlib function would run interpreted
-//! forever (no LIR → never submitted to the JIT worker) — a silent runtime
-//! regression, so LIR is serialized too, with its `doc`/`syntax` Rc fields
-//! skipped (they are already `None` after the cross-thread conversion in
-//! `sendable_from_template`; JIT never reads them).
+// audited: 2026-09-09
+//! Serialize the compiled standard library to a content-addressed file, so a
+//! later process deserializes it instead of running the front end again.
+//! docs/impl/stdlib-cache.md
 
 use crate::compiler::Bytecode;
 use crate::signals::Signal;
@@ -57,13 +33,10 @@ fn payload_hash(bytes: &[u8]) -> u64 {
 
 /// The on-disk form of a compiled module's entry `Bytecode`.
 ///
-/// The whole `Bytecode` is wrapped as a synthetic entry `ClosureTemplate` and
-/// serialized through the send module's template path (`serialize_templates`),
-/// which deep-copies the entry constant pool (it may contain live closure
-/// instances — stdlib's `init_stdlib` result closure, `map`, etc.), the
-/// nested-lambda blueprints, their LIR, and the region-release tables. The
-/// two extra fields below (`signal_projection`, `format_version`) don't exist
-/// on `ClosureTemplate`, so they ride alongside.
+/// The `Bytecode` rides in `entry`, wrapped as a synthetic `ClosureTemplate` so
+/// the send module's template path carries the whole cyclic closure graph
+/// (docs/impl/stdlib-cache.md). Every other field is one a `ClosureTemplate`
+/// has nowhere to hold.
 #[derive(Serialize, Deserialize)]
 pub struct StoredBytecode {
     pub format_version: u32,
@@ -149,22 +122,14 @@ fn cache_key(stdlib_source: &str) -> Option<String> {
     use std::hash::{Hash, Hasher};
     let mut hasher = std::collections::hash_map::DefaultHasher::new();
     stdlib_source.hash(&mut hasher);
-    // The binary itself, not its version string. Two builds of one version
-    // compile stdlib differently the moment the emitter or a pass changes, and
-    // a key that cannot see the rebuild hands every `Runtime::new()` — the one
-    // in each test included — bytecode the previous binary produced. A test
-    // failure would then stop implicating the branch that caused it.
-    //
-    // It also covers what the primitive-table identity below cannot see: the
-    // ids `prim_id_of` appends outside the canonical tables (trait methods,
-    // FFI callbacks) are minted by the running binary, not listed in it.
+    // The binary itself, not its version string: two builds of one version
+    // compile stdlib differently the moment a pass changes
+    // (docs/impl/stdlib-cache.md).
     build_identity()?.hash(&mut hasher);
     FORMAT_VERSION.hash(&mut hasher);
-    // A serialized native-fn immediate carries a `prim_id`, which is only
-    // valid against the exact primitive table that minted it. Mix the table
-    // identity in so a prim addition/removal/reorder (a different elle
-    // binary) invalidates the cache instead of deserializing a foreign id
-    // into `panic!("unknown prim id")`.
+    // A serialized native-fn immediate carries a `prim_id`, valid only against
+    // the table that minted it; without this a foreign id reaches
+    // `panic!("unknown prim id")` instead of a miss.
     crate::primitives::registration::hash_prim_table_identity(&mut hasher);
     Some(format!("{:016x}.bin", hasher.finish()))
 }
@@ -230,11 +195,8 @@ pub fn try_store(
             let path = dir.join(key);
             let mut file = payload_hash(&bytes).to_le_bytes().to_vec();
             file.extend_from_slice(&bytes);
-            // Write beside the target and rename over it. Two elle processes
-            // starting at once is ordinary, and writing the final path directly
-            // lets one read the other's half-written file — or edits the inode a
-            // reader already holds open. A rename is one atomic step within a
-            // directory, so the name refers to a complete file or the old one.
+            // Never write `path` directly: another process starting right now
+            // would read a half-written file (docs/impl/stdlib-cache.md).
             if let Err(e) = store_atomically(&dir, &path, &file) {
                 eprintln!("[stdlib-cache] write failed: {e}");
                 return;
@@ -265,16 +227,13 @@ fn store_atomically(
         .map_err(|e| std::io::Error::other(e.to_string()))
 }
 
-/// Remove every cache file in `dir` except `keep`.
+/// Remove every cache file in `dir` except `keep` — the megabytes each earlier
+/// build's key orphans (docs/impl/stdlib-cache.md).
 ///
-/// The key follows the binary, so every rebuild mints a new one and orphans
-/// the file the last one wrote. At ~16 MB each, a day of rebuilds fills a
-/// directory nobody thinks to look at. Run after the rename, never before: a
-/// store that fails must leave the directory as it found it, still holding a
-/// file some other process may be about to read.
-///
-/// A removal that fails is ignored. It is disk hygiene, not correctness, and
-/// the next store tries again.
+/// Call this after the rename, never before: a store that fails must leave the
+/// directory as it found it, still holding a file some other process may be
+/// about to read. A removal that fails is ignored; it is disk hygiene, not
+/// correctness, and the next store tries again.
 fn prune_superseded(dir: &std::path::Path, keep: &std::path::Path) {
     let Ok(entries) = std::fs::read_dir(dir) else {
         return;
@@ -289,12 +248,10 @@ fn prune_superseded(dir: &std::path::Path, keep: &std::path::Path) {
 
 /// Serialize compiled stdlib bytecode into the cache format.
 ///
-/// The whole `Bytecode` is wrapped as a synthetic entry `ClosureTemplate`
-/// (arity `Exact(0)` — the entry runs as a thunk) and serialized through the
-/// send module's template path. This handles everything uniformly: the entry
-/// constant pool (which may hold live closure instances), the nested-lambda
-/// blueprints, their LIR (so the JIT keeps working after reload), and the
-/// region-release tables.
+/// The `Bytecode` becomes a synthetic entry `ClosureTemplate` of arity
+/// `Exact(0)` — it runs as a thunk — and goes through the send module's
+/// template path, which carries the entry pool, the nested-lambda blueprints,
+/// their LIR and the region-release tables uniformly.
 pub fn store_bytecode(
     bytecode: &Bytecode,
     vm: &mut crate::vm::VM,
@@ -391,501 +348,4 @@ pub fn load_bytecode(
 }
 
 #[cfg(test)]
-mod tests {
-    use super::*;
-    use crate::pipeline::compile_file;
-    use crate::primitives::module_init::StdlibSource;
-    use crate::runtime::Runtime;
-
-    /// Compile a snippet through the full pipeline, then assert that
-    /// store→load round-trips to an equivalent `Bytecode` (equal instructions
-    /// and constants, closures rebuilt, LIR preserved).
-    #[test]
-    fn bytecode_roundtrip_preserves_lir_and_closures() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        // Its own directory: this test writes a cache and must not read, write,
-        // or be read by whatever else the suite is running beside it.
-        let mut rt = Runtime::with_stdlib_cache(StdlibCache::Dir(dir.path().to_path_buf()));
-        let (result, loaded) = {
-            let (vm, symbols, cctx) = rt.parts();
-            let src = r#"
-(defn helper [x] (+ x 1))
-(+ (helper 1) (helper 2))
-"#;
-            let result = compile_file(src, symbols, cctx, "<test>").expect("compiles");
-            let bc = &result.bytecode;
-            assert!(!bc.instructions.is_empty());
-
-            let stored = store_bytecode(bc, vm, symbols, cctx).expect("stores");
-            let stored_names = stored.names.len();
-            let bytes = bincode::serialize(&stored).expect("serializes");
-            let decoded: StoredBytecode = bincode::deserialize(&bytes).expect("deserializes");
-            assert_eq!(
-                decoded.names.len(),
-                stored_names,
-                "the spelling table survives bincode; without it every reloaded \
-                 symbol prints as #<symbol:hash>"
-            );
-            let loaded = load_bytecode(decoded, vm, symbols, cctx).expect("loads");
-            assert_eq!(loaded.instructions, bc.instructions, "instructions equal");
-            assert_eq!(loaded.signal, bc.signal);
-            assert_eq!(loaded.child_protos.len(), bc.child_protos.len());
-            // The constant pool is byte-identical on the scalar prefix; closure
-            // constants are NEW heap instances after reload (pointer-equal
-            // comparison would spuriously fail), so compare scalar kinds/counts.
-            assert_eq!(
-                bc.constants.len(),
-                loaded.constants.len(),
-                "same number of constants"
-            );
-            for (a, b) in bc.constants.iter().zip(&loaded.constants) {
-                assert_eq!(a.is_closure(), b.is_closure(), "closure-ness preserved");
-                assert_eq!(a.is_heap(), b.is_heap(), "heap-ness preserved");
-            }
-            // LIR must survive (JIT depends on it) and closures must be rebuilt.
-            for (orig, reloaded) in bc.child_protos.iter().zip(&loaded.child_protos) {
-                assert_eq!(
-                    orig.lir_function.is_some(),
-                    reloaded.lir_function.is_some(),
-                    "LIR presence preserved"
-                );
-            }
-            let _ = vm;
-            (result.bytecode, loaded)
-        };
-        // Both bytecodes must execute to the same result.
-        let run = |bc: &crate::compiler::Bytecode| -> i64 {
-            let (vm, _symbols, cctx) = rt.parts();
-            vm.execute_scheduled(bc, cctx)
-                .expect("runs")
-                .as_int()
-                .expect("result is an int")
-        };
-        let mut run = run;
-        let r_orig = run(&result);
-        let r_loaded = run(&loaded);
-        assert_eq!(r_orig, r_loaded, "original and reloaded bytecode agree");
-        eprintln!("roundtrip ok: {r_orig} == {r_loaded}");
-    }
-
-    /// The key must follow the binary, not its version string: two builds of
-    /// one version compile stdlib differently the moment a pass changes. This
-    /// pins that the identity is read from the executable — an edit that put a
-    /// constant back would leave every rebuild sharing one key, and every test
-    /// that boots a runtime reading bytecode the previous binary produced.
-    ///
-    /// Characterization, not a failing-first regression: "a different build
-    /// yields a different key" is observable only across builds.
-    #[test]
-    fn the_build_identity_is_read_from_the_running_executable() {
-        let (len, _mtime) = build_identity().expect("the test binary can identify itself");
-        let exe = std::env::current_exe().expect("current_exe");
-        let meta = std::fs::metadata(&exe).expect("exe metadata");
-        assert_eq!(
-            len,
-            meta.len(),
-            "the identity must be the binary's own size"
-        );
-        assert!(len > 0, "a zero-length identity separates nothing");
-    }
-
-    /// Two runtimes over one cache directory: the first compiles stdlib and
-    /// writes the cache, the second must load from it — and must still have a
-    /// working stdlib afterwards.
-    #[test]
-    fn second_runtime_on_a_shared_cache_dir_loads_stdlib_from_it() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let cache = StdlibCache::Dir(dir.path().to_path_buf());
-
-        let mut a = Runtime::with_stdlib_cache(cache.clone());
-        assert_eq!(
-            a.stdlib_source(),
-            StdlibSource::Compiled,
-            "the first runtime meets an empty directory, so it must compile"
-        );
-
-        let mut b = Runtime::with_stdlib_cache(cache);
-        assert_eq!(
-            b.stdlib_source(),
-            StdlibSource::Cache,
-            "the second runtime must load what the first wrote; a cache that \
-             silently never hits still yields a working runtime, so behaviour \
-             alone cannot tell the two apart"
-        );
-
-        // Working stdlib on both sides. Functional check only — timing is
-        // asserted in the release-mode boot benchmark instead (debug builds
-        // skew it).
-        let probe = |rt: &mut Runtime| -> crate::value::Value {
-            use crate::pipeline::compile_file_repl;
-            let (vm, symbols, cctx) = rt.parts();
-            let src = "(map (fn [x] (* x 2)) (quote (1 2 3)))";
-            let result = compile_file_repl(src, symbols, cctx, "<probe>").expect("probe compiles");
-            vm.execute_scheduled(&result.0.bytecode, cctx)
-                .expect("probe runs")
-        };
-        let _ = probe(&mut a);
-        let _ = probe(&mut b);
-    }
-
-    /// A `sys/spawn` worker runs `init_stdlib` on its own thread, so it reads
-    /// and writes a cache of its own. It must use the directory its parent was
-    /// given: a worker that falls back to the process-wide one writes megabytes
-    /// into a place nobody named, which is the leak the construction parameter
-    /// exists to close.
-    #[test]
-    fn a_spawned_worker_caches_where_its_parent_was_told_to() {
-        use crate::pipeline::compile_file_repl;
-
-        let dir = tempfile::tempdir().expect("tempdir");
-        let mut rt = Runtime::with_stdlib_cache(StdlibCache::Dir(dir.path().to_path_buf()));
-
-        // The parent's own store already filled the directory, and the worker
-        // writes the same key. Clear it, so anything present afterwards can
-        // only have been written by the worker.
-        let entries = |p: &std::path::Path| -> usize {
-            std::fs::read_dir(p).expect("read cache dir").count()
-        };
-        for entry in std::fs::read_dir(dir.path()).expect("read cache dir") {
-            std::fs::remove_file(entry.expect("entry").path()).expect("clear");
-        }
-        assert_eq!(entries(dir.path()), 0, "cleared");
-
-        let (vm, symbols, cctx) = rt.parts();
-        let result =
-            compile_file_repl("(sys/join (sys/spawn (fn [] 1)))", symbols, cctx, "<spawn>")
-                .expect("spawn form compiles");
-        vm.execute_scheduled(&result.0.bytecode, cctx)
-            .expect("spawn runs");
-
-        assert_eq!(
-            entries(dir.path()),
-            1,
-            "the worker must cache into the directory its parent was given, \
-             not the process-wide one"
-        );
-    }
-
-    /// A cache file is bytes on a disk any process can write. `bincode` reports
-    /// only that the bytes *decoded*, never that they are the bytes this binary
-    /// wrote — so a flipped byte reaches the VM as instructions, and a deeper
-    /// one is absorbed into stdlib and reported as a hit. Every corruption must
-    /// come out as a miss: the cache is an optimization, and a full compile is
-    /// always available.
-    #[test]
-    fn a_corrupt_cache_file_is_a_miss_not_a_panic_or_a_silent_edit() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let cache = StdlibCache::Dir(dir.path().to_path_buf());
-
-        // Seed a good cache, then keep its bytes to restore between rounds.
-        drop(Runtime::with_stdlib_cache(cache.clone()));
-        let path = std::fs::read_dir(dir.path())
-            .expect("read cache dir")
-            .next()
-            .expect("the seeding runtime wrote a cache file")
-            .expect("entry")
-            .path();
-        let good = std::fs::read(&path).expect("read cache file");
-        assert!(
-            good.len() > 8192,
-            "cache file too small to corrupt meaningfully"
-        );
-
-        // Two offsets, because there are two failure modes: a shallow one
-        // lands in the decoded structure and reaches the VM as instructions, a
-        // deep one lands in a payload and passes unnoticed. More offsets in the
-        // shallow band would repeat a mode rather than add one, and each round
-        // costs a runtime boot.
-        for offset in [64usize, good.len() / 2] {
-            let mut bad = good.clone();
-            for byte in &mut bad[offset..offset + 8] {
-                *byte = 0xFF;
-            }
-            std::fs::write(&path, &bad).expect("write corrupt cache");
-
-            let rt = Runtime::with_stdlib_cache(cache.clone());
-            assert_eq!(
-                rt.stdlib_source(),
-                StdlibSource::Compiled,
-                "eight corrupt bytes at offset {offset} must be a miss"
-            );
-        }
-    }
-
-    /// Falling back is half the job. A rejected file that stays on disk is
-    /// rejected again by every later start, so one bad write costs the cache
-    /// permanently — the recompile it forces is invisible, because a working
-    /// runtime is what a miss produces too.
-    #[test]
-    fn a_rejected_cache_file_is_replaced_not_left_to_win() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let cache = StdlibCache::Dir(dir.path().to_path_buf());
-
-        drop(Runtime::with_stdlib_cache(cache.clone()));
-        let path = std::fs::read_dir(dir.path())
-            .expect("read cache dir")
-            .next()
-            .expect("the seeding runtime wrote a cache file")
-            .expect("entry")
-            .path();
-        std::fs::write(&path, b"not a cache file").expect("plant a rejected file");
-
-        let first = Runtime::with_stdlib_cache(cache.clone());
-        assert_eq!(
-            first.stdlib_source(),
-            StdlibSource::Compiled,
-            "the planted file must be rejected"
-        );
-        drop(first);
-
-        let second = Runtime::with_stdlib_cache(cache);
-        assert_eq!(
-            second.stdlib_source(),
-            StdlibSource::Cache,
-            "the runtime that rejected the file must also replace it, or every \
-             later start pays the full compile again"
-        );
-    }
-
-    /// Two elle processes starting at once is an ordinary event, and a store
-    /// that writes the final path directly lets one of them read the other's
-    /// half-written file. The store must land whole or not at all.
-    ///
-    /// A held descriptor is the observable: writing the path in place edits the
-    /// inode the reader already has, while a rename leaves that inode alone and
-    /// swings the name to a new one.
-    #[test]
-    fn a_store_replaces_the_cache_file_instead_of_rewriting_it() {
-        use std::io::Read;
-
-        let dir = tempfile::tempdir().expect("tempdir");
-        let cache = StdlibCache::Dir(dir.path().to_path_buf());
-
-        drop(Runtime::with_stdlib_cache(cache.clone()));
-        let path = std::fs::read_dir(dir.path())
-            .expect("read cache dir")
-            .next()
-            .expect("the seeding runtime wrote a cache file")
-            .expect("entry")
-            .path();
-
-        // Unreadable, so the next runtime rejects it and must store over it.
-        const SENTINEL: &[u8] = b"the inode a reader already holds";
-        std::fs::write(&path, SENTINEL).expect("plant a rejected file");
-        let mut held = std::fs::File::open(&path).expect("hold the old inode open");
-
-        drop(Runtime::with_stdlib_cache(cache));
-
-        let mut seen = Vec::new();
-        held.read_to_end(&mut seen)
-            .expect("read through the held fd");
-        // Compared as a bool: the rewritten file is megabytes, and dumping it
-        // into the failure would bury the one fact that matters.
-        assert!(
-            seen == SENTINEL,
-            "the store must rename a complete file into place; rewriting the \
-             path edits the inode another process is already reading — the \
-             held descriptor saw {} bytes, not the {}-byte sentinel",
-            seen.len(),
-            SENTINEL.len()
-        );
-    }
-
-    /// Every key that stops being current orphans a file, and the key follows
-    /// the binary — so an ordinary day of rebuilds leaves one 16 MB file per
-    /// build, forever, in a directory nobody thinks to look at.
-    #[test]
-    fn a_store_prunes_the_files_its_key_supersedes() {
-        let dir = tempfile::tempdir().expect("tempdir");
-        let cache = StdlibCache::Dir(dir.path().to_path_buf());
-
-        // Two files under keys this binary will never mint again.
-        for name in ["deadbeefdeadbeef.bin", "0123456789abcdef.bin"] {
-            std::fs::write(dir.path().join(name), b"an earlier build's cache")
-                .expect("plant a superseded file");
-        }
-
-        drop(Runtime::with_stdlib_cache(cache));
-
-        let left: Vec<_> = std::fs::read_dir(dir.path())
-            .expect("read cache dir")
-            .map(|e| e.expect("entry").file_name())
-            .collect();
-        assert_eq!(
-            left.len(),
-            1,
-            "a store must leave only the file it just wrote; found {left:?}"
-        );
-    }
-
-    /// The registry a cache hit restores must be the registry the stdlib
-    /// compile recorded. It drives an HIR rewrite in every later compile, so a
-    /// snapshot that drops entries makes the cached path compile user code
-    /// differently from the compiled path — and caching is on by default, so
-    /// the first run of a program and every run after it disagree.
-    ///
-    /// The counter-factual: a snapshot that omits the templates whose body is a
-    /// `let` — the shape whose clone needed the defining arena — loses seven of
-    /// the stdlib's thirty-six, and the assertion names them.
-    #[test]
-    fn the_stored_inline_registry_keeps_every_template_the_compile_recorded() {
-        use std::collections::BTreeSet;
-
-        fn names(
-            reg: &crate::hir::typeinfer::FnInlineRegistry,
-            symbols: &crate::symbol::SymbolTable,
-        ) -> BTreeSet<String> {
-            reg.by_name
-                .keys()
-                .map(|n| symbols.name(*n).unwrap_or("?").to_string())
-                .collect()
-        }
-
-        let dir = tempfile::tempdir().expect("tempdir");
-        let mut rt = Runtime::with_stdlib_cache(StdlibCache::Dir(dir.path().to_path_buf()));
-        let (_vm, symbols, cctx) = rt.parts();
-
-        let (recorded, stored) = {
-            let (_, fn_inline) = cctx.compile_registries_mut();
-            (names(fn_inline, symbols), fn_inline.to_stored(symbols))
-        };
-        assert!(
-            !recorded.is_empty(),
-            "the stdlib compile must record cross-unit inline templates, or \
-             this test proves nothing about what the snapshot keeps"
-        );
-
-        let mut restored = crate::hir::typeinfer::FnInlineRegistry::default();
-        restored.restore(stored, symbols);
-
-        let survived = names(&restored, symbols);
-        let lost: Vec<_> = recorded.difference(&survived).collect();
-        assert!(
-            lost.is_empty(),
-            "a cache hit must inline what a stdlib compile inlines; these \
-             templates do not survive the snapshot: {lost:?}"
-        );
-    }
-
-    /// Two boot paths, one rewrite. `fn/cfg-label` is a stdlib `defn` whose body
-    /// is a `let` — the shape a registry that could not carry bindings had to
-    /// drop — and fusing it into a `map` splices that body into the emitted
-    /// loop, leaving no call to it at all.
-    ///
-    /// The trap: the two paths' instruction streams are not byte-identical even
-    /// for `(+ 1 2)`, because bytecode operands carry ids from a process-global
-    /// mint counter that the stdlib compile advances and a cache hit does not.
-    /// So the claim is asserted where it lives — in the rewritten HIR — with the
-    /// stream lengths as corroboration.
-    #[test]
-    fn a_cache_hit_inlines_the_stdlib_bodies_a_stdlib_compile_inlines() {
-        use crate::hir::{BindingArena, Hir, HirKind};
-        use crate::pipeline::{compile_file_repl, compile_file_to_fhir};
-
-        const SRC: &str = r#"(map fn/cfg-label [{:name "a"} {:name "b"}])"#;
-        const INLINED: &str = "fn/cfg-label";
-
-        fn calls_named(
-            h: &Hir,
-            arena: &BindingArena,
-            symbols: &crate::symbol::SymbolTable,
-            want: &str,
-        ) -> usize {
-            let mut n = 0;
-            if let HirKind::Call { func, .. } = &h.kind {
-                if let HirKind::Var(b) = &func.kind {
-                    n += usize::from(symbols.name(arena.get(*b).name) == Some(want));
-                }
-            }
-            h.for_each_child(|c| n += calls_named(c, arena, symbols, want));
-            n
-        }
-
-        fn probe(rt: &mut Runtime) -> (usize, usize) {
-            let (_vm, symbols, cctx) = rt.parts();
-            let (hir, arena) =
-                compile_file_to_fhir(SRC, symbols, cctx, "<parity>").expect("compiles to HIR");
-            let calls = calls_named(&hir, &arena, symbols, INLINED);
-            let len = compile_file_repl(SRC, symbols, cctx, "<parity>")
-                .expect("compiles")
-                .0
-                .bytecode
-                .instructions
-                .len();
-            (calls, len)
-        }
-
-        let dir = tempfile::tempdir().expect("tempdir");
-        let cache = StdlibCache::Dir(dir.path().to_path_buf());
-
-        let mut compiled = Runtime::with_stdlib_cache(cache.clone());
-        assert_eq!(compiled.stdlib_source(), StdlibSource::Compiled);
-        let (compiled_calls, compiled_len) = probe(&mut compiled);
-        drop(compiled);
-
-        let mut hit = Runtime::with_stdlib_cache(cache);
-        assert_eq!(hit.stdlib_source(), StdlibSource::Cache);
-        let (cached_calls, cached_len) = probe(&mut hit);
-
-        assert_eq!(
-            compiled_calls, 0,
-            "the stdlib compile records the fragment, so the `map` fuses and \
-             the call to `{INLINED}` is gone"
-        );
-        assert_eq!(
-            cached_calls, 0,
-            "a cache hit must fuse the same call; a registry that dropped the \
-             fragment leaves the un-fused call behind"
-        );
-        assert_eq!(
-            compiled_len, cached_len,
-            "and the two paths must emit the same amount of code for it"
-        );
-    }
-
-    /// `ClosureTemplate.origin` does not cross the cache — a restore rebuilds
-    /// templates from cached bytecode, not from the LIR the emitter set it on
-    /// — and `(meta/origin f)`, its only reader, reports a closure's source
-    /// location from it. So a stdlib closure has an origin on the compiled
-    /// path and none on the cached one. Nothing in the tree depends on that;
-    /// it is pinned here rather than left to be rediscovered as a surprise.
-    ///
-    /// The second half is what bounds it: a closure the hit runtime compiles
-    /// itself still knows where it came from, so the loss stays with the values
-    /// the cache restored and does not reach user code.
-    #[test]
-    fn a_cached_stdlib_closure_has_no_origin_but_user_code_keeps_its_own() {
-        fn origin_is_nil(rt: &mut Runtime, src: &str) -> bool {
-            use crate::pipeline::compile_file_repl;
-            let (vm, symbols, cctx) = rt.parts();
-            let result = compile_file_repl(src, symbols, cctx, "<origin>").expect("compiles");
-            vm.execute_scheduled(&result.0.bytecode, cctx)
-                .expect("runs")
-                .is_nil()
-        }
-
-        let dir = tempfile::tempdir().expect("tempdir");
-        let cache = StdlibCache::Dir(dir.path().to_path_buf());
-
-        let mut compiled = Runtime::with_stdlib_cache(cache.clone());
-        assert_eq!(compiled.stdlib_source(), StdlibSource::Compiled);
-        assert!(
-            !origin_is_nil(&mut compiled, "(meta/origin map)"),
-            "a compiled stdlib closure carries its origin span"
-        );
-        drop(compiled);
-
-        let mut hit = Runtime::with_stdlib_cache(cache);
-        assert_eq!(hit.stdlib_source(), StdlibSource::Cache);
-        assert!(
-            origin_is_nil(&mut hit, "(meta/origin map)"),
-            "the cache does not carry the origin span, so a restored closure \
-             has none — a known difference between the two paths"
-        );
-        assert!(
-            !origin_is_nil(&mut hit, "(meta/origin (fn [] nil))"),
-            "the loss must stay with the restored values: a closure this \
-             runtime compiled itself still knows where it came from"
-        );
-    }
-}
+mod tests;
