+++ src/compiler/stdlib_cache/tests/mod.rs
@@ -1,0 +1,7 @@
+// audited: 2026-09-09
+// The cache has two subjects and one test file each: what the serialized form
+// carries across, and how the file on disk behaves.
+// docs/impl/stdlib-cache.md
+
+mod file;
+mod format;
